requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.29.1
frozenlist==1.3.0

httplib2==0.20.4
idna==3.3
kiwisolver==1.3.2

multidict==6.0.2

opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.0.1
pip==22.0.4
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7

python-dateutil==2.8.2
pytz==2021.3
pyzbar==0.1.9
requests==2.27.1
rsa==4.8
setuptools==57.0.0
six==1.16.0
urllib3==1.26.8
wheel==0.36.2
yarl==1.7.2

dotenvy~=0.2.0